phase3/phase2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "sqlmodel>=0.0.14",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "apscheduler>=3.10.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.5.0",
    "python-dotenv>=1.0.0",
]
